requirements.txt
paddlepaddle==2.2.0
typed-argument-parser
numpy